chore:  creating the useCalculator hook
